feature/ai/chatbot

See merge request s12-webmobile1-sub1/S12P11E104!104

+++ be/src/main/java/com/be/domain/chatbot/service/ChatService.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import java.util.List;
 
 public interface ChatService {
-    StartResponse chatStart(Integer user_id, StartRequest request) throws JsonProcessingException;
+    StartResponse chatStart(Long user_id, StartRequest request) throws JsonProcessingException;
     String chat(String sessionId, String userMessage) throws JsonProcessingException;
     String generateSummary(List<String> chatHistory);
 }
